Exception Handler added

+++ src/main/java/com/assignment/newsportal/service/PostService.java
@@ -1,6 +1,8 @@
 package com.assignment.newsportal.service;
 
 import com.assignment.newsportal.Exception.DuplicateDataException;
+import com.assignment.newsportal.Exception.InvalidRequestException;
+import com.assignment.newsportal.Exception.MissingDetailException;
 import com.assignment.newsportal.Exception.NotFoundException;
 import com.assignment.newsportal.dto.request.PostDTO;
 import com.assignment.newsportal.entity.*;
@@ -37,7 +39,7 @@ public class PostService {
     PostHashtagRepo postHashtagRepo;
 
 
-    private final Logger logger= LoggerFactory.getLogger(PostService.class);
+    private final Logger logger = LoggerFactory.getLogger(PostService.class);
 
 //    public Post publish(Post post){
 //        post.setDraft(false); post.setActive(true);
@@ -52,7 +54,7 @@ public class PostService {
         return postRepo.getActivePosts(pageable);
     }
 
-    public void unpublish(Post post,Long userId) {
+    public void unpublish(Post post, Long userId) {
 
 //        Post post=postRepo.findByPostId(postId).orElse(null);
 //        if(post==null||!post.getActive()){
@@ -68,86 +70,59 @@ public class PostService {
 
     }
 
-//    public Post upvote(Long postId) {
-//        Post post=postRepo.findByPostId(postId).orElse(null);
-//        if(post==null||!post.getActive())
-//            throw new NotFoundException("Post with id "+postId+" doesn't exist");
-//
-//        //Long upvotes=post.getUpvotes();
-//        post.setUpvotes(post.getUpvotes()+1);
-//        return postRepo.save(post);
-//    }
-
-//    public Post downvote(Long postId) {
-//        Post post=postRepo.findByPostId(postId).orElse(null);
-//        if(post==null||!post.getActive())
-//            throw new NotFoundException("Post with id "+postId+" doesn't exist");
-//
-//        //Long upvotes=post.getUpvotes();
-//        post.setDownvotes((post.getDownvotes()+1));
-//        return postRepo.save(post);
-//    }
-
-//    public Post update(Post post, PostDTO postDTO) {
-//        post.setTitle(postDTO.getTitle());
-//        post.setBody(postDTO.getBody());
-////        post.setUpdatedBy(post.getUserId());
-//
-//       return postRepo.save(post);
-//    }
     @Transactional
     public Post vote(Long postId, Long val, Long userId) {
         Post post = postRepo.findByPostId(postId).orElse(null);
         if (post == null || !post.getActive())
             throw new NotFoundException("Post with id " + postId + " doesn't exist");
-        VoteStatus voteStatus= voteStatusRepo.find(userId,postId).orElse(null);
-        if(voteStatus==null) {
-            voteStatus=new VoteStatus(userId,postId,true,false,false);
+        VoteStatus voteStatus = voteStatusRepo.find(userId, postId).orElse(null);
+
+        if((val!=1)&&(val!=-1))
+        throw new InvalidRequestException("Options are 1 for upvote and -1 for downvote.");
+
+        if (voteStatus == null) {
+            voteStatus = new VoteStatus(userId, postId, true, false, false);
 
             if (val == 1) {
                 voteStatus.setUpvoted(true);
                 voteStatus.setDownvoted(false);
-                post.setUpvotes(post.getUpvotes()+1);
+                post.setUpvotes(post.getUpvotes() + 1);
                 voteStatusRepo.save(voteStatus);
 
             }
-              else if(val==-1) {
+            else if (val == -1) {
 
                 voteStatus.setUpvoted(false);
                 voteStatus.setDownvoted(true);
-                post.setDownvotes(post.getDownvotes()+1);
+                post.setDownvotes(post.getDownvotes() + 1);
                 voteStatusRepo.save(voteStatus);
 
 
-              }
-              else{
-                  throw new IllegalArgumentException("Options are 1 for upvote and -1 for downvote.");
             }
-        }
-        else if(voteStatus!=null){
+//            else {
+//                throw new InvalidRequestException("Options are 1 for upvote and -1 for downvote.");
+//            }
+        } else if (voteStatus != null) {
 
-            if(val==1&&voteStatus.getUpvoted()) {
-                logger.error("You have already upvoted.");
+            if (val == 1 && voteStatus.getUpvoted()) {
+                throw new InvalidRequestException("You have already upvoted.");
 
-            }
-            else if(val==1&&(voteStatus.getDownvoted())){
+            } else if (val == 1 && (voteStatus.getDownvoted())) {
                 voteStatus.setUpvoted(true);
                 voteStatus.setDownvoted(false);
-                post.setUpvotes(post.getUpvotes()+1);
-                post.setDownvotes(post.getDownvotes()-1);
+                post.setUpvotes(post.getUpvotes() + 1);
+                post.setDownvotes(post.getDownvotes() - 1);
                 voteStatusRepo.save(voteStatus);
                 logger.info("Changed downvote to upvote");
 
-            }
-            else if((val==-1)&&(voteStatus.getDownvoted())){
-                logger.error("You have already downvoted.");
+            } else if ((val == -1) && (voteStatus.getDownvoted())) {
+                throw new InvalidRequestException("You have already downvoted.");
 
-            }
-            else if((val==-1)&&(voteStatus.getUpvoted())){
+            } else if ((val == -1) && (voteStatus.getUpvoted())) {
                 voteStatus.setUpvoted(false);
                 voteStatus.setDownvoted(true);
-                post.setUpvotes(post.getUpvotes()-1);
-                post.setDownvotes(post.getDownvotes()+1);
+                post.setUpvotes(post.getUpvotes() - 1);
+                post.setDownvotes(post.getDownvotes() + 1);
                 voteStatusRepo.save(voteStatus);
                 logger.info("Changed upvote to downvote.");
 
@@ -156,63 +131,75 @@ public class PostService {
 
         }
 
-           return postRepo.save(post);
+        return postRepo.save(post);
 
 
     }
 
     @Transactional
     public Post publish(PostDTO postDTO, Long userId) {
-        Set<String> hashtags=postDTO.getHashtags();
-        List<Hashtag> hashtagList= new ArrayList<>();
-         for(String h: hashtags){
-             Hashtag hashtag= hashtagRepo.findByHashtag(h).orElse(null);
-             if((hashtag==null)||(hashtag.getActive()==false))
-                 throw new NotFoundException("Hashtag "+h+" does not exist.");
-             hashtagList.add(hashtag);
-         }
 
-         Post post=new Post(postDTO.getTitle(),postDTO.getBody(),hashtags,0L,0L,false,true,userId);
-         post.setCreatedBy(userId);
-         post.setUpdatedBy(userId);
-         post=postRepo.save(post);
+        if(postDTO.getTitle().equals("")){
+            throw new MissingDetailException("Post title is incomplete");
+        }
+        Set<String> hashtags = postDTO.getHashtags();
+        List<Hashtag> hashtagList = new ArrayList<>();
+        for (String h : hashtags) {
+            Hashtag hashtag = hashtagRepo.findByHashtag(h).orElse(null);
+            if ((hashtag == null) || (hashtag.getActive() == false))
+                throw new NotFoundException("Hashtag " + h + " does not exist.");
+            hashtagList.add(hashtag);
+        }
 
-         Long postId=post.getPostId();
+        Post post = new Post(postDTO.getTitle(), postDTO.getBody(), hashtags, 0L, 0L, false, true, userId);
+        post.setCreatedBy(userId);
+        post.setUpdatedBy(userId);
+        post = postRepo.save(post);
 
-         for(Hashtag hashtag:hashtagList){
-             PostHashtagMap postHashtagMap= new PostHashtagMap(postId,hashtag.getHashtagId(),hashtag.getHashtag(),hashtag.getTopicId(),userId);
-             postHashtagMap.setCreatedBy(userId);
-             postHashtagMap.setUpdatedBy(userId);
-             postHashtagRepo.save(postHashtagMap);
-         }
+        Long postId = post.getPostId();
 
-         return post;
+        for (Hashtag hashtag : hashtagList) {
+            PostHashtagMap postHashtagMap = new PostHashtagMap(postId, hashtag.getHashtagId(), hashtag.getHashtag(), hashtag.getTopicId(), userId);
+            postHashtagMap.setCreatedBy(userId);
+            postHashtagMap.setUpdatedBy(userId);
+            postHashtagRepo.save(postHashtagMap);
+        }
+
+
+        return post;
     }
 
     @Transactional
-    public Post update(Post post, PostDTO postDTO) {
+    public Post update(Long postId, PostDTO postDTO) {
 
+//        Long postId = post.getPostId();
 
-        Long postId=post.getPostId();
-        Long userId=post.getUserId();
+        if(postDTO.getTitle().equals("")){
+            throw new MissingDetailException("Post title is incomplete");
+        }
+        Post post=postRepo.findByPostId(postId).orElse(null);
+        if(post==null||!post.getActive()){
+            throw new NotFoundException("Post with id "+postId+" not found.");
+        }
+        Long userId = post.getUserId();
         postHashtagRepo.deleteByPostId(postId);
-        Set<String> hashtags=postDTO.getHashtags();
-        List<Hashtag> hashtagList= new ArrayList<>();
-        for(String h: hashtags){
-            Hashtag hashtag= hashtagRepo.findByHashtag(h).orElse(null);
-            if((hashtag==null)||(hashtag.getActive()==false))
-                throw new NotFoundException("Hashtag "+h+" does not exist.");
+        Set<String> hashtags = postDTO.getHashtags();
+        List<Hashtag> hashtagList = new ArrayList<>();
+        for (String h : hashtags) {
+            Hashtag hashtag = hashtagRepo.findByHashtag(h).orElse(null);
+            if ((hashtag == null) || (hashtag.getActive() == false))
+                throw new NotFoundException("Hashtag " + h + " does not exist.");
             hashtagList.add(hashtag);
         }
 
         post.setTitle(postDTO.getTitle());
         post.setBody(postDTO.getBody());
         post.setHashtags(hashtags);
-        post=postRepo.save(post);
+        post = postRepo.save(post);
 
 
-        for(Hashtag hashtag:hashtagList){
-            PostHashtagMap postHashtagMap= new PostHashtagMap(postId,hashtag.getHashtagId(),hashtag.getHashtag(),hashtag.getTopicId(),userId);
+        for (Hashtag hashtag : hashtagList) {
+            PostHashtagMap postHashtagMap = new PostHashtagMap(postId, hashtag.getHashtagId(), hashtag.getHashtag(), hashtag.getTopicId(), userId);
             postHashtagMap.setCreatedBy(userId);
             postHashtagMap.setUpdatedBy(userId);
             postHashtagRepo.save(postHashtagMap);
@@ -220,5 +207,15 @@ public class PostService {
 
         return post;
     }
+
+    public Page<Post> search(String searchVal, Pageable pageable) {
+        if (searchVal.equals(""))
+            throw new MissingDetailException("Search value not provided");
+        if (searchVal.startsWith("#")) {
+            return postRepo.findByHashtag(searchVal, pageable);
+        }
+        return postRepo.findByTitle(searchVal, pageable);
+    }
+
 }
